type cast

+++ tests/test_stream_manager.c
@@ -1326,7 +1326,7 @@ TEST_CASE(localhost_integ_h2_sm_acquire_stream_stress) {
     };
     ASSERT_SUCCESS(s_tester_init(&options));
     size_t num_to_acquire = 500 * 100;
-    ASSERT_SUCCESS(s_sm_stream_acquiring(num_to_acquire));
+    ASSERT_SUCCESS(s_sm_stream_acquiring((int)num_to_acquire));
     ASSERT_SUCCESS(s_wait_on_streams_completed_count(num_to_acquire));
     ASSERT_UINT_EQUALS(s_tester.acquiring_stream_errors, 0);
     ASSERT_UINT_EQUALS(s_tester.stream_200_count, num_to_acquire);
